Criação do MessageService para facilitar o uso de mensagens Correções de imports não usados Correções de indentação Refatoração de testes Refatoração da página de cadastrar evento

+++ aplicacao/contest/src/main/java/ufc/quixada/npi/contest/repository/EventoRepository.java
@@ -9,6 +9,6 @@ import ufc.quixada.npi.contest.model.Evento;
 
 @Repository
 @Transactional
-public interface EventoRepository extends CrudRepository<Evento, Integer>{
+public interface EventoRepository extends CrudRepository<Evento, Long>{
 	
 }
